HSPC-4700
When adding a class or style to a field group that has no data.layout, and when the style change requires the field group to be rendered in a grid-like way (such as setting the background-color), the style can appear to have no effect. This standardizes the display property of field groups that use the default layout system so it is more consistent with field groups that do have a layout.