HSFW-170
Updates gulp tasks to use the new approach to versioning. When built in a nightly kit, the version info will be pulled from the build file. Otherwise it is set to "internal" and the build date and time.

Also removes the distinction between "core" and "app" version properties, and puts them into the runtime `app.config` instead of a separate `version.json` file.

`iscVersionApi` is maintained for backwards compatibility with any app that uses it, but now it just queries the `app.config` version info (asynchronously) instead of `version.json`.