HSFW-171
Fixes issue of the `iscConfirmation` modal not working when "production" is set to true in `app.config`. This is because the "production" flag disables all debug-related activity, and this modal currently uses a debug operation to obtain the zf-modal scope and show the modal. This refactors this behavior to use the Foundation API `publish` method instead.